+++ nextstep-di/src/test/java/nextstep/di/factory/circularreference/CircularReferenceController.java
@@ -1,0 +1,12 @@
+package nextstep.di.factory.circularreference;
+
+import nextstep.annotation.Inject;
+
+public class CircularReferenceController {
+    private CircularReferenceController circularReferenceController;
+
+    @Inject
+    public CircularReferenceController(CircularReferenceController circularReferenceController) {
+        this.circularReferenceController = circularReferenceController;
+    }
+}
