--- nextstep-di/src/test/java/nextstep/di/factory/example/NotBean.java
@@ -1,4 +1,0 @@
-package nextstep.di.factory.example;
-
-public class NotBean {
-}
