--- nextstep-di/src/test/java/nextstep/di/factory/example/RecursiveController.java
@@ -1,12 +1,0 @@
-package nextstep.di.factory.example;
-
-import nextstep.annotation.Inject;
-
-public class RecursiveController {
-    private RecursiveController recursiveController;
-
-    @Inject
-    public RecursiveController(RecursiveController recursiveController) {
-        this.recursiveController = recursiveController;
-    }
-}
