+++ nextstep-di/src/test/java/nextstep/di/factory/notbean/InjectNotBean.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.notbean;
 
 import nextstep.annotation.Inject;
