+++ nextstep-di/src/test/java/nextstep/di/factory/notbean/NotBean.java
@@ -1,0 +1,4 @@
+package nextstep.di.factory.notbean;
+
+public class NotBean {
+}
